server/requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
bcrypt==4.3.0
cffi==2.0.0
click==8.2.1
cryptography==45.0.7
databases==0.8.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2==2.9.10
psycopg2-binary==2.9.7
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.0.3
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
